Carve strict functions out to a separate module
